requirements.txt
streamlit
scikit-learn
joblib
pdfplumber
python-docx
numpy
pandas
